requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.1
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
pycparser==2.22
PyJWT==2.8.0
requests==2.31.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.0
stripe==9.7.0
typing_extensions==4.11.0
urllib3==2.2.1
